+++ NINA/ViewModel/Imaging/AnchorableSnapshotVM.cs
@@ -34,6 +34,9 @@ using static NINA.Equipment.Model.CaptureSequence;
 using NINA.Equipment.Equipment;
 using NINA.WPF.Base.Interfaces.ViewModel;
 using NINA.WPF.Base.ViewModel;
+using NINA.Astrometry;
+using NINA.WPF.Base.Behaviors;
+using System.ComponentModel;
 
 namespace NINA.ViewModel.Imaging {
 
@@ -95,6 +98,7 @@ namespace NINA.ViewModel.Imaging {
             profileService.ProfileChanged += (object sender, EventArgs e) => {
                 SnapFilter = profileService.ActiveProfile.FilterWheelSettings.FilterWheelFilters?.FirstOrDefault(x => x.Name == profileService.ActiveProfile.SnapShotControlSettings.Filter?.Name);
             };
+            SubSampleRectangleMoveCommand = new RelayCommand(SubSampleRectangleMove);
         }
 
         /// <summary>
@@ -215,6 +219,87 @@ namespace NINA.ViewModel.Imaging {
             }
         }
 
+        private ObservableRectangle subSampleRectangle;
+
+        public ObservableRectangle SubSampleRectangle {
+            get => subSampleRectangle;
+            set {
+                if (subSampleRectangle != null) {
+                    subSampleRectangle.PropertyChanged -= SubSampleRectangle_PropertyChangedSizeValidation;
+                }
+                subSampleRectangle = value;
+                if (subSampleRectangle != null) {
+                    subSampleRectangle.PropertyChanged += SubSampleRectangle_PropertyChangedSizeValidation;
+                }
+                RaisePropertyChanged();
+            }
+        }
+
+        private void SubSampleRectangleMove(object obj) {
+            var dragResult = (DragResult)obj;
+            if (SubSampleRectangle != null) {
+                var mode = dragResult.Mode;
+                var delta = dragResult.Delta;
+                if (mode == DragMode.Move) {
+                    if (SubSampleRectangle.X + delta.X < 0) {
+                        delta.X = 0;
+                    } else if (SubSampleRectangle.X + SubSampleRectangle.Width + delta.X > CameraInfo.XSize) {
+                        delta.X = (int)(CameraInfo.XSize - SubSampleRectangle.Width - SubSampleRectangle.X);
+                    }
+
+                    if (SubSampleRectangle.Y + delta.Y < 0) {
+                        delta.Y = 0;
+                    } else if (SubSampleRectangle.Y + SubSampleRectangle.Height + delta.Y > CameraInfo.YSize) {
+                        delta.Y = (int)(CameraInfo.YSize - SubSampleRectangle.Height - SubSampleRectangle.Y);
+                    }
+
+                    var x = (int)(SubSampleRectangle.X + delta.X);
+                    var y = (int)(SubSampleRectangle.Y + delta.Y);
+                    SubSampleRectangle = new ObservableRectangle(x, y, SubSampleRectangle.Width, SubSampleRectangle.Height);
+                } else {
+                    var x = (int)SubSampleRectangle.X;
+                    var y = (int)SubSampleRectangle.Y;
+                    var width = (int)SubSampleRectangle.Width;
+                    var height = (int)SubSampleRectangle.Height;
+
+                    if (mode == DragMode.Resize_Top_Left) {
+                        x += (int)delta.X;
+                        y += (int)delta.Y;
+                        width -= (int)delta.X;
+                        height -= (int)delta.Y;
+                    } else if (mode == DragMode.Resize_Top_Right) {
+                        y += (int)delta.Y;
+                        width += (int)delta.X;
+                        height -= (int)delta.Y;
+                    } else if (mode == DragMode.Resize_Bottom_Left) {
+                        x += (int)delta.X;
+                        width -= (int)delta.X;
+                        height += (int)delta.Y;
+                    } else if (mode == DragMode.Resize_Left) {
+                        x += (int)delta.X;
+                        width -= (int)delta.X;
+                    } else if (mode == DragMode.Resize_Right) {
+                        width += (int)delta.X;
+                    } else if (mode == DragMode.Resize_Top) {
+                        y += (int)delta.Y;
+                        height -= (int)delta.Y;
+                    } else if (mode == DragMode.Resize_Bottom) {
+                        height += (int)delta.Y;
+                    } else {
+                        width += (int)delta.X;
+                        height += (int)delta.Y;
+                    }
+                    SubSampleRectangle = new ObservableRectangle(x, y, width, height);
+                }
+            }
+        }
+
+        private void SubSampleRectangle_PropertyChangedSizeValidation(object sender, PropertyChangedEventArgs e) {
+            SubSampleRectangleMove(new DragResult() { Delta = new System.Windows.Vector(), Mode = DragMode.Move });
+        }
+
+        public ICommand SubSampleRectangleMoveCommand { get; private set; }
+
         public IAsyncCommand StartLiveViewCommand { get; private set; }
 
         public ApplicationStatus Status {
@@ -270,6 +355,7 @@ namespace NINA.ViewModel.Imaging {
                 do {
                     var seq = new CaptureSequence(SnapExposureDuration, ImageTypes.SNAPSHOT, SnapFilter, SnapBin, 1);
                     seq.EnableSubSample = SnapSubSample;
+                    seq.SubSambleRectangle = SubSampleRectangle;
                     seq.Gain = SnapGain;
 
                     var exposureData = await imagingMediator.CaptureImage(seq, _captureImageToken.Token, progress);
@@ -308,6 +394,12 @@ namespace NINA.ViewModel.Imaging {
 
         public void UpdateDeviceInfo(CameraInfo cameraStatus) {
             CameraInfo = cameraStatus;
+            if (cameraInfo.Connected && SubSampleRectangle == null) {
+                SubSampleRectangle = new ObservableRectangle(cameraInfo.SubSampleX, cameraInfo.SubSampleY, cameraInfo.SubSampleWidth, cameraInfo.SubSampleHeight);
+            } else if (!cameraInfo.Connected && SubSampleRectangle != null) {
+                SnapSubSample = false;
+                SubSampleRectangle = null;
+            }
         }
     }
 }