+++ NINA/ViewModel/ImagingVM.cs
@@ -260,9 +260,6 @@ namespace NINA.ViewModel {
                     return data;
                 } finally {
                     progress.Report(new ApplicationStatus() { Status = string.Empty });
-                    if (ImageControl.ShowSubSampler) {
-                        ImageControl.ShowSubSampler = false;
-                    }
                 }
             });
         }